+++ source/main/utils/setupLogging.js
@@ -1,8 +1,10 @@
+import fs from 'fs';
 import path from 'path';
 import log from 'electron-log';
 import moment from 'moment';
 import ensureDirectoryExists from './ensureDirectoryExists';
-import { pubLogsFolderPath, APP_NAME } from '../config';
+import { pubLogsFolderPath, appLogsFolderPath, APP_NAME } from '../config';
+import { isFileNameWithTimestamp } from '../../common/fileName';
 
 const isTest = process.env.NODE_ENV === 'test';
 
@@ -19,4 +21,18 @@ export const setupLogging = () => {
     const formattedDate = moment.utc(msg.date).format('YYYY-MM-DDTHH:mm:ss.0SSS');
     return `[${formattedDate}Z] [${msg.level}] ${msg.data}`;
   };
+
+  // Removes existing compressed logs
+  fs.readdir(appLogsFolderPath, (err, files) => {
+    files
+      .filter(isFileNameWithTimestamp())
+      .forEach((logFileName) => {
+        const logFile = path.join(appLogsFolderPath, logFileName);
+        try {
+          fs.unlinkSync(logFile);
+        } catch (error) {
+          console.error(`Compressed log file "${logFile}" deletion failed: ${error}`);
+        }
+      });
+  });
 };